requirements.txt
feedparser>=6.0.0
requests>=2.28.0
python-dateutil>=2.8.0
beautifulsoup4>=4.12.0
